+++ contracts/interfaces/IFrankTreasury.sol
@@ -37,7 +37,7 @@ interface IFrankTreasury {
 
     function setSlippage (uint256 _slippage) external;
 
-    function setStrategy(uint16[2] memory _DISTRIBUTION_BONDED_JOE, uint16[3] memory _DISTRIBUTION_REINVESTMENTS, uint16 _PROPORTION_REINVESTMENTS, address _LIQUIDITY_POOL) external;
+    function setStrategy(uint256[2] memory _DISTRIBUTION_BONDED_JOE, uint256[3] memory _DISTRIBUTION_REINVESTMENTS, uint256 _PROPORTION_REINVESTMENTS, address _LIQUIDITY_POOL) external;
 
     function distribute() external;
